move eval to right place
